+++ NEWS.md
@@ -2,6 +2,7 @@
 
 # v0.1.3
 
+* New function `detect()`.
 * New function `dictionary_sugar()`.
 * New function `dictionary_cast()` (moved from `mlr3`).
 * It is now asserted that the return value of Dictionaries is an R6 object.